fix: 19425: Audit Foundation / Data XTS tests
