+++ platform-sdk/swirlds-merkledb/build.gradle.kts
@@ -25,33 +25,10 @@ testModuleInfo {
     requires("com.swirlds.common.test.fixtures")
     requires("com.swirlds.config.extensions.test.fixtures")
     requires("com.swirlds.merkledb.test.fixtures")
+    requires("org.apache.commons.lang3")
+    requires("org.apache.logging.log4j.core")
     requires("org.hiero.base.utility.test.fixtures")
     requires("org.hiero.consensus.model")
-    requires("org.junit.jupiter.api")
-    requires("org.junit.jupiter.params")
-    requires("org.mockito")
-}
-
-timingSensitiveModuleInfo {
-    requires("com.google.common")
-    requires("com.swirlds.base")
-    requires("com.swirlds.common")
-    requires("com.swirlds.common.test.fixtures")
-    requires("com.swirlds.config.api")
-    requires("com.swirlds.config.extensions")
-    requires("com.swirlds.config.extensions.test.fixtures")
-    requires("com.swirlds.merkledb")
-    requires("com.swirlds.merkledb.test.fixtures")
-    requires("com.swirlds.metrics.api")
-    requires("com.swirlds.virtualmap")
-    requires("org.hiero.base.concurrent")
-    requires("org.hiero.base.crypto")
-    requires("org.hiero.base.utility")
-    requires("org.hiero.consensus.model")
-    requires("org.apache.commons.lang3")
-    requires("org.apache.logging.log4j")
-    requires("org.apache.logging.log4j.core")
-    requires("org.eclipse.collections.impl")
     requires("org.junit.jupiter.api")
     requires("org.junit.jupiter.params")
     requires("org.mockito")
