+++ platform-sdk/swirlds-merkledb/src/test/java/com/swirlds/merkledb/MerkleDbDataSourceTest.java
@@ -45,6 +45,7 @@ import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Random;
+import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
@@ -837,6 +838,61 @@ class MerkleDbDataSourceTest {
                 copy.close();
             }
         });
+    }
+
+    @ParameterizedTest
+    @EnumSource(TestType.class)
+    void closeWhileFlushingTest(final TestType testType) throws IOException, InterruptedException {
+        final Path dbPath = testDirectory.resolve("merkledb-closeWhileFlushingTest-" + testType);
+        final MerkleDbDataSource dataSource = testType.dataType().createDataSource(dbPath, "vm", 1000, 0, false, false);
+
+        final int count = 20;
+        final List<VirtualKey> keys = new ArrayList<>(count);
+        for (int i = 0; i < count; i++) {
+            keys.add(testType.dataType().createVirtualLongKey(i));
+        }
+        final List<ExampleByteArrayVirtualValue> values = new ArrayList<>(count);
+        for (int i = 0; i < count; i++) {
+            values.add(testType.dataType().createVirtualValue(i + 1));
+        }
+
+        final CountDownLatch updateStarted = new CountDownLatch(1);
+        final Thread closeThread = new Thread(() -> {
+            try {
+                updateStarted.await();
+                Thread.sleep(new Random().nextInt(100));
+                dataSource.close();
+            } catch (Exception z) {
+                // Print and ignore
+                z.printStackTrace(System.err);
+            }
+        });
+        closeThread.start();
+
+        final KeySerializer keySerializer = testType.dataType().getKeySerializer();
+        final ValueSerializer valueSerializer = testType.dataType().getValueSerializer();
+
+        updateStarted.countDown();
+        for (int i = 0; i < 10; i++) {
+            final int k = i;
+            try {
+                dataSource.saveRecords(
+                        count - 1,
+                        2 * count - 2,
+                        IntStream.range(0, count).mapToObj(j -> new VirtualHashRecord(k + j, hash(k + j + 1))),
+                        IntStream.range(count - 1, count)
+                                .mapToObj(j -> new VirtualLeafRecord<>(k + j, keys.get(k), values.get((k + j) % count)))
+                                .map(r -> r.toBytes(keySerializer, valueSerializer)),
+                        Stream.empty(),
+                        true);
+            } catch (Exception z) {
+                // Print and ignore
+                z.printStackTrace(System.err);
+                break;
+            }
+        }
+
+        closeThread.join();
     }
 
     // =================================================================================================================
