+++ platform-sdk/swirlds-merkledb/src/test/java/com/swirlds/merkledb/files/CloseFlushTest.java
@@ -33,6 +33,8 @@ import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Stream;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.core.config.Configurator;
+import org.hiero.base.constructable.ClassConstructorPair;
+import org.hiero.base.constructable.ConstructableRegistry;
 import org.hiero.base.crypto.Hash;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
@@ -64,7 +66,7 @@ public class CloseFlushTest {
 
     @Test
     public void closeFlushTest() throws Exception {
-        final int count = 100000;
+        final int count = 10000;
         final ExecutorService exec = Executors.newSingleThreadExecutor();
         final AtomicReference<Exception> exception = new AtomicReference<>();
         for (int j = 0; j < 100; j++) {
@@ -98,7 +100,7 @@ public class CloseFlushTest {
             final VirtualMap<VirtualKey, ExampleByteArrayVirtualValue> lastMap = map;
             final Future<?> job = exec.submit(() -> {
                 try {
-                    Thread.sleep(new Random().nextInt(500));
+                    Thread.sleep(new Random().nextInt(100));
                     lastMap.release();
                 } catch (final Exception z) {
                     throw new RuntimeException(z);
@@ -123,8 +125,8 @@ public class CloseFlushTest {
         private AtomicReference<Exception> exceptionSink = null;
 
         // Provided for deserialization
-        public CustomDataSourceBuilder(final @NonNull Configuration configuration) {
-            super(configuration);
+        public CustomDataSourceBuilder() {
+            super(CONFIGURATION);
         }
 
         public CustomDataSourceBuilder(
