+++ platform-sdk/swirlds-merkledb/src/test/java/com/swirlds/merkledb/files/DataFileCollectionCompactionTest.java
@@ -510,7 +510,8 @@ class DataFileCollectionCompactionTest {
                 // during merge,
                 // it depends on where pauseCompaction() happens inside
                 // compactFiles() above
-                assertTrue(List.of(2, 3).contains(store.getAllCompletedFiles().size()));
+                assertTrue(List.of(2, 3).contains(store.getAllCompletedFiles().size()),
+                        "Unexpected files after compaction: " + store.getAllCompletedFiles());
             } catch (final Exception e) {
                 throw new RuntimeException(e);
             } finally {
