--- platform-sdk/swirlds-merkledb/src/timingSensitive/resources/log4j2-test.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright 2016-2022 Hedera Hashgraph, LLC
-  ~
-  ~ This software is the confidential and proprietary information of
-  ~ Hedera Hashgraph, LLC. ("Confidential Information"). You shall not
-  ~ disclose such Confidential Information and shall use it only in
-  ~ accordance with the terms of the license agreement you entered into
-  ~ with Hedera Hashgraph.
-  ~
-  ~ HEDERA HASHGRAPH MAKES NO REPRESENTATIONS OR WARRANTIES ABOUT THE SUITABILITY OF
-  ~ THE SOFTWARE, EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED
-  ~ TO THE IMPLIED WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
-  ~ PARTICULAR PURPOSE, OR NON-INFRINGEMENT. HEDERA HASHGRAPH SHALL NOT BE LIABLE FOR
-  ~ ANY DAMAGES SUFFERED BY LICENSEE AS A RESULT OF USING, MODIFYING OR
-  ~ DISTRIBUTING THIS SOFTWARE OR ITS DERIVATIVES.
-  -->
-
-<Configuration status="WARN">
-	<Appenders>
-		<Console name="Console" target="SYSTEM_OUT">
-			<PatternLayout pattern="%style{%t}{bright} %highlight{%-1level} %style{%logger{0}}{cyan} %msg%n"
-						   disableAnsi="false"/>
-		</Console>
-	</Appenders>
-	<Loggers>
-		<Root level="trace">
-			<AppenderRef ref="Console"/>
-		</Root>
-	</Loggers>
-</Configuration>
