+++ platform-sdk/swirlds-virtualmap/src/test/resources/virtualRootNode_ver1/rootNode.bin
