+++ platform-sdk/swirlds-virtualmap/src/test/resources/virtualRootNode_ver2/rootNode.bin
